Add files via upload
Am an aspiring to become a Data Scientist and am learning various Data Toolkits and exploring the world of Data Science. 
This Project is what I have done as a final project for my Executive PG Program in Data Science and Business Analytics. 
This involves Analysis of Stocks of companies listed in NYSE belonging to various industries such as Aviation, Finance, Healthcare and Technology. The project contains steps of Data Cleaning, Data Visualisation, Exploratory Data Analysis, Time Series Analysis, Beta Value and Expected Returns calculation using CAPM and finally Monte_Carlo Simulation through which the Best Portfolio is selected. 
The Best Portfolio is found and the same is suggested to the Investor.